make useLike hook
