+++ client/src/component/FeedModal.tsx
@@ -4,7 +4,7 @@ import { Feed } from "../types/Feed";
 import LikeButton from "./LikeButton";
 import { useSelector } from "react-redux";
 import { RootState } from "../store/rootStore";
-import axios from "axios";
+import useLike from "../hooks/useLike";
 export default function FeedModal({
     card,
     onClose = () => {}
@@ -14,20 +14,16 @@ export default function FeedModal({
 }){
     const [images, setImages] = useState<string[]>([])
     const [currentIndex, setCurrentIndex] = useState(0)
-    const [likeCount, setLikeCount] = useState(card.like_count)
-    const [isLiked, setIsLiked] = useState(false)
     const writeDate = new Date(card.created_at).toDateString()
     const user = useSelector((state: RootState) => state.auth.user)
+    const user_id = user?.id
+    const {isLiked, likeCount, toggleLike} = useLike(user_id, card.id)
     
     useEffect(()=>{
         card.images.forEach(async (img)=>{
             await getImageUrl(img)
         })
     }, [card.images])
-
-    useEffect(()=>{
-        checkLike()
-    }, [])
 
     function getImageUrl(image: string){
         const fullUrl  = `${process.env.REACT_APP_SUPABASE_URL}/storage/v1/object/public/feed-images/${card.user_id}/${card.id}/${image}`
@@ -41,68 +37,6 @@ export default function FeedModal({
 
     function goPrev(){
         setCurrentIndex(prev => (prev - 1 + images.length) % images.length)
-    }
-
-    async function checkLike(){
-        try {
-            const user_id = user?.id
-            const response = await axios.post('/api/feed/checkLike', { user_id: user_id, feed_id: card.id})
-            if(response.status === 200){
-                setIsLiked(response.data.liked)
-            }
-        } catch(error){
-            console.log(error)
-        }
-    }
-
-    function clickLike(){
-        if(!isLiked){
-            setLike()
-        } else {
-            cancelLike()
-        }
-    }
-
-    async function getLikeCount(){
-            try {
-                const response = await axios.post('/api/feed/getLike', { feed_id: card.id})
-                if(response.status === 200){
-                    console.log(`count: ${response.data}`)
-                    const count = response.data
-                    setLikeCount(count)
-                }
-            } catch(error){
-                console.log(error)
-            }
-    }
-
-    async function setLike(){
-        if(isLiked) return;
-        const user_id = user?.id
-        try {
-            const response = await axios.post('/api/feed/like', {user_id: user_id, feed_id: card.id})
-            if(response.status === 200){
-                getLikeCount()
-                setIsLiked(true)
-            }
-        } catch(error){
-            console.log(error)
-        }
-    }
-
-    async function cancelLike(){
-        if(!isLiked) return;
-        
-        const user_id = user?.id
-        try {
-            const response = await axios.post('/api/feed/unlike', { user_id: user_id, feed_id: card.id})
-            if(response.status === 200){
-                getLikeCount()
-                setIsLiked(false)
-            }
-        } catch(error){
-            console.log(error)
-        }
     }
 
     return (
@@ -140,7 +74,7 @@ export default function FeedModal({
                     </div>
                     <div className={styles.information}>
                         <div className={styles.likeZone}>
-                            <LikeButton isLiked={isLiked} width="20px" height="20px" onClick={clickLike}/>
+                            <LikeButton isLiked={isLiked} width="20px" height="20px" onClick={toggleLike}/>
                             <p>{likeCount}</p>
                         </div>
                         <p className={styles.id}>{card.userId}</p>
